Added bikedrums from DB

+++ Models/BikeContainer.cs
@@ -7,8 +7,8 @@ namespace AndroidBicycleInfo
 	{
 		[PrimaryKey]
 		public int id { get; set; }
-		public float lat { get; set; }
-		public float lon { get; set; }
+		public double lat { get; set; }
+		public double lon { get; set; }
 
 		[Indexed]
 		public int street_id { get; set; }
